+++ dev/java-test.js
@@ -5,7 +5,7 @@ const util = require("util");
 
 const broker = new ServiceBroker({
 	nodeID: "node-js",
-	transporter: "redis://192.168.51.72:6379",
+	transporter: "redis://localhost:6379",
 	//logLevel: "debug",
 });
 
@@ -14,8 +14,8 @@ broker.loadService("./examples/math.service.js");
 broker.start()
 	.then(() => broker.repl())
 	.then(() => {
-		setInterval(() => {
+		/*setInterval(() => {
 			broker.sendPing().then(res => broker.logger.info("Ping result:", res));
 
-		}, 2000);
+		}, 2000);*/
 	});